# Samat_Check/FBA/portal/fba/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "FBA POS API",
    "version": "1.0"
  },
  "host": "FBAappuv01.fortebank.com:8082",
  "paths": {
...
    "/point": {
      "put": {
        "description": "Изменение торговой точки",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Point"
        ],
        "operationId": "PointPUT",
        "parameters": [
          {
            "description": "JSON для изменения модели торговой точки",
            "name": "bodyjson",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Point"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON ответ на изменение торговой точки",
            "schema": {
              "$ref": "#/definitions/Point"
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "post": {
        "description": "Создание торговой точки",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Point"
        ],
        "operationId": "PointPOST",
        "parameters": [
          {
            "description": "JSON для создания торговой точки",
            "name": "bodyjson",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Point"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON созданная модель торговой точки",
            "schema": {
              "$ref": "#/definitions/Point"
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/point/{id}": {
      "get": {
        "description": "Получение торговой точки по ID",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Point"
        ],
        "operationId": "PointGETByID",
        "parameters": [
          {
            "type": "string \u003cuuid4\u003e",
            "description": "идентификатор торговой точки",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "JSON модель торговой точки",
            "schema": {
              "$ref": "#/definitions/Point"
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "delete": {
        "description": "Удаление торговой точки по ID",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Point"
        ],
        "operationId": "PointDELETE",
        "parameters": [
          {
            "type": "string \u003cuuid4\u003e",
            "description": "идентификатор торговой точки",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное удаление",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/point/{id}/authorize": {
      "get": {
        "description": "Получение информации по клиенту OAuth связанному с торговой точкой",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Point"
        ],
        "operationId": "PointGetOauthClient",
        "parameters": [
          {
            "type": "string \u003cuuid4\u003e",
            "description": "идентификатор торговой точки",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "JSON модель клиента в OAuth",
            "schema": {
              "$ref": "#/definitions/OAuth2Client"
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/points": {
      "get": {
        "description": "Получение списка городов",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Point"
        ],
        "operationId": "PointGET",
        "parameters": [
          {
            "type": "int",
            "description": "номер страницы пагинатора",
            "name": "page",
            "in": "query"
          },
          {
            "type": "int",
            "description": "количество записей на странице",
            "name": "size",
            "in": "query"
          },
          {
            "type": "string \u003cuuid4\u003e",
            "description": "идентификатор компании",
            "name": "companyid",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "JSON массив городов",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Point"
              }
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/points/count": {
      "get": {
        "description": "Получение количества торговых точек",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Point"
        ],
        "operationId": "PointCount",
        "responses": {
          "200": {
            "description": "JSON с количеством торговых точек",
            "schema": {
              "$ref": "#/definitions/Count"
            }
          },
          "500": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "City": {
      "description": "City справочник городов",
      "type": "object",
      "properties": {
        "ID": {
          "type": "string",
          "format": "uuid4"
        },
        "Name": {
          "type": "string"
        }
      },
      "x-go-package": "fba/dbmodel"
    },
    "Company": {
      "description": "Company компании",
      "type": "object",
      "properties": {
        "Address": {
          "type": "string"
        },
        "Bin": {
          "type": "string"
        },
        "Contact": {
          "type": "string"
        },
        "ID": {
          "type": "string",
          "format": "uuid4"
        },
        "Name": {
          "type": "string"
        },
        "Staff": {
          "$ref": "#/definitions/Staff"
        },
        "StaffID": {
          "type": "string",
          "format": "uuid4"
        }
      },
      "x-go-package": "fba/dbmodel"
    },
    "Count": {
      "description": "modelRowsCount структура для количества записей под swagger",
      "type": "object",
      "properties": {
        "count": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Count"
        }
      },
      "x-go-name": "modelRowsCount",
      "x-go-package": "fba/model"
    },
...
    "Error": {
      "description": "Error Ошибка",
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-package": "fba/model"
    },
...
    "OAuth2Client": {
      "description": "Client oauth2 client",
      "type": "object",
      "properties": {
        "CreatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "DeletedAt": {
          "type": "string",
          "format": "date-time"
        },
        "Domain": {
          "type": "string"
        },
        "ID": {
          "type": "string",
          "format": "uuid4"
        },
        "Secret": {
          "type": "string"
        },
        "UpdatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "UserID": {
          "type": "string",
          "format": "uuid4"
        }
      },
      "x-go-name": "Client",
      "x-go-package": "fba/http"
    },
...
    "Point": {
      "description": "Point точки продаж",
      "type": "object",
      "properties": {
        "Address": {
          "type": "string"
        },
        "City": {
          "$ref": "#/definitions/City"
        },
        "CityID": {
          "type": "string",
          "format": "uuid4"
        },
        "Company": {
          "$ref": "#/definitions/Company"
        },
        "CompanyID": {
          "type": "string",
          "format": "uuid4"
        },
        "ID": {
          "type": "string",
          "format": "uuid4"
        },
        "Name": {
          "type": "string"
        },
        "Staff": {
          "$ref": "#/definitions/Staff"
        },
        "StaffID": {
          "type": "string",
          "format": "uuid4"
        }
      },
      "x-go-package": "fba/dbmodel"
    },
...
    "Role": {
      "description": "Role права пользователей(staff)",
      "type": "object",
      "properties": {
        "ID": {
          "type": "string",
          "format": "uuid4"
        },
        "Name": {
          "type": "string"
        }
      },
      "x-go-package": "fba/dbmodel"
    },
    "Staff": {
      "description": "Staff пользователь",
      "type": "object",
      "properties": {
        "ID": {
          "type": "string",
          "format": "uuid4"
        },
        "Name": {
          "type": "string"
        },
        "ParentID": {
          "type": "string",
          "format": "uuid4"
        },
        "Role": {
          "$ref": "#/definitions/Role"
        },
        "RoleID": {
          "type": "string",
          "format": "uuid4"
        }
      },
      "x-go-package": "fba/dbmodel"
    },
...
  },
  "security": [
    {
      "api_key": []
    }
  ]
}